# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Caching proxy Service API",
    "version": "1.0.0",
    "description": "API for the caching_proxy Cache Service, a Node.js/TypeScript application acting as a caching middleware layer.\nIt proxies requests to a backend service, caches GET responses based on path and user role,\nand provides administrative endpoints for monitoring and cache management.\n"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server (adjust port based on .env)"
    }
  ],
...
  "paths": {
...
    "/keys": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List Cache Keys",
        "description": "Retrieves a list of all keys currently stored in the cache (Redis/KeyDB).\nKeys are grouped by the role identifier (`anonymous`, `authenticated`, `admin`, etc.) derived from the cache key structure (`cache:METHOD:ROLE:PATH`).\n**Warning:** This operation uses `SCAN` and can be resource-intensive on the cache instance if the number of keys is very large. Use with caution in production.\n",
        "operationId": "listCacheKeys",
        "responses": {
          "200": {
            "description": "A list of cache keys grouped by role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheKeysList"
                }
              }
            }
          },
          "500": {
            "description": "Error listing cache keys.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/invalidate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Manually Invalidate Cache",
        "description": "Manually invalidates (deletes) cache entries matching a specified Redis key pattern.\nSupports wildcards (`*`). Use carefully, especially with broad patterns.\nIf the pattern targets a specific path without specifying a role (e.g., `cache:GET:*:/api/users/123`), the service automatically invalidates entries for *all* roles matching that path.\nIf the pattern includes a specific role (e.g., `cache:GET:anonymous:/api/users/123`), only entries for that role are targeted.\n",
        "operationId": "invalidateCache",
        "requestBody": {
          "required": true,
          "description": "The pattern to use for cache invalidation.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvalidateRequest"
              },
              "example": {
                "pattern": "cache:GET:*:/api/articles/some-article-slug"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invalidation process completed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Missing or invalid 'pattern' in the request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Missing or invalid \"pattern\" in request body"
                }
              }
            }
          },
          "500": {
            "description": "Error during cache invalidation process.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CacheKeysList": {
        "type": "object",
        "properties": {
          "keyCount": {
            "type": "integer",
            "format": "int32",
            "description": "Total number of keys currently in the cache."
          },
          "keysByRole": {
            "type": "object",
            "description": "Cache keys grouped by the role identifier found within the key.",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "cache:GET:anonymous:/api/articles/public"
              }
            }
          },
          "keys": {
            "type": "array",
            "description": "A flat list of all cache keys found.",
            "items": {
              "type": "string",
              "example": "cache:GET:authenticated:/api/user/profile"
            }
          }
        },
        "required": [
          "keyCount",
          "keysByRole",
          "keys"
        ]
      },
      "InvalidateRequest": {
        "type": "object",
        "properties": {
          "pattern": {
            "type": "string",
            "description": "The Redis key pattern to use for invalidation (e.g., `cache:GET:*:/api/users/*`). Wildcards (`*`) are supported.",
            "example": "cache:GET:*:*"
          }
        },
        "required": [
          "pattern"
        ]
      },
      "InvalidateResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Confirmation message indicating the pattern processed.",
            "example": "Invalidation process completed for pattern: cache:GET:*:/api/articles/123"
          },
          "deletedCount": {
            "type": "integer",
            "format": "int32",
            "description": "The number of cache keys deleted based on the provided pattern."
          }
        },
        "required": [
          "message",
          "deletedCount"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "A high-level error category or message.",
            "example": "Internal Server Error"
          },
          "message": {
            "type": "string",
            "description": "More detailed error message (potentially omitted in production).",
            "example": "Failed to connect to Redis"
          }
        },
        "required": [
          "error"
        ]
      }
    },
...
  }
}